Reworked TestApp, splitting beacons into multiple screens & added push
