+++ TestApp/android/app/build.gradle
@@ -137,6 +137,7 @@ android {
 }
 
 dependencies {
+    compile project(':mobile-center-push')
     compile project(':mobile-center-analytics')
     compile project(':mobile-center-crashes')
     compile fileTree(dir: "libs", include: ["*.jar"])
