+++ TestApp/android/app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
+    <string moduleConfig="true" name="mobileCenterPush_whenToEnablePush">ALWAYS_SEND</string>
     <string moduleConfig="true" name="mobileCenterAnalytics_whenToEnableAnalytics">ALWAYS_SEND</string>
     <string moduleConfig="true" name="mobileCenterCrashes_whenToSendCrashes">ASK_JAVASCRIPT</string>
     <string name="app_name">TestApp</string>
