+++ TestApp/package.json
@@ -1,25 +1,27 @@
 {
-	"name": "TestApp",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"start": "node node_modules/react-native/local-cli/cli.js start",
-		"test": "jest"
-	},
-	"dependencies": {
-		"react": "15.4.2",
-		"react-native": "0.40.0",
- 		"mobile-center-link-scripts": "../mobile-center-link-scripts",
-		"mobile-center-analytics": "../mobile-center-analytics",
-		"mobile-center-crashes": "../mobile-center-crashes"
-	},
-	"devDependencies": {
-		"babel-jest": "18.0.0",
-		"babel-preset-react-native": "1.9.1",
-		"jest": "18.1.0",
-		"react-test-renderer": "15.4.2"
-	},
-	"jest": {
-		"preset": "react-native"
-	}
+  "name": "TestApp",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "test": "jest"
+  },
+  "dependencies": {
+    "mobile-center-analytics": "../mobile-center-analytics",
+    "mobile-center-crashes": "../mobile-center-crashes",
+    "mobile-center-push": "../mobile-center-push",
+    "mobile-center-link-scripts": "../mobile-center-link-scripts",
+    "react": "15.4.2",
+    "react-native": "0.40.0",
+    "react-navigation": "^1.0.0-beta.11"
+  },
+  "devDependencies": {
+    "babel-jest": "18.0.0",
+    "babel-preset-react-native": "1.9.1",
+    "jest": "18.1.0",
+    "react-test-renderer": "15.4.2"
+  },
+  "jest": {
+    "preset": "react-native"
+  }
 }
