+++ TestApp/update-link-to-mobile-center.sh
@@ -3,6 +3,8 @@
 echo 'Unlinking...'
 react-native unlink mobile-center-crashes
 react-native unlink mobile-center-analytics
+react-native unlink mobile-center-push
 echo 'Linking...'
 react-native link mobile-center-crashes
 react-native link mobile-center-analytics
+react-native link mobile-center-push
