+++ TestApp/update-npm-packages.sh
@@ -3,4 +3,5 @@ echo 'Updating our npm packages...'
 rm -rf node_modules/mobile-center-link-scripts
 rm -rf node_modules/mobile-center-crashes
 rm -rf node_modules/mobile-center-analytics
+rm -rf node_modules/mobile-center-push
 npm install
